Find Occurrences in File
Program made in intel 8086 asssembly (made for MASM).
This program looks for occurences in a file and shows the previous and the next words between the searched word. The name of the file is passed by console parameter.
To run this, is necessary to use the Dosbox simulator and the MASM assembler.